fix(tmux): correct #1793 chunking claims, fix UTF-8 split, cut chunk-delay overhead
Addresses review findings on PR #1819 (fix/1793-canonical-chunk):

- The sendKeysChunkedToTarget comment claimed the fix works because it keeps
  each write under the Linux tty canonical-mode buffer (MAX_CANON=4096).
  That reasoning is mechanically invalid: canonical-mode buffering is keyed
  on the line, not the individual write() call, so splitting a newline-free
  payload into N smaller writes still occupies the same bytes in the kernel
  line buffer before Enter arrives. Rewrote the comment to state this
  honestly, drop the unsubstantiated "measured-safe" claim (the reported
  Codex CLI repro was never run against this change — host policy forbids
  running live agent CLI sessions here), and describe chunking + delay as a
  heuristic defense-in-depth mitigation, not a proven fix for the reported
  mechanism.

- splitIntoChunks now backs the hard-split fallback cut point off to the
  nearest UTF-8 rune boundary instead of slicing at a raw byte offset, so a
  multibyte character can no longer be split across two tmux argv writes.
  Added TestSplitIntoChunks_PreservesUTF8RuneBoundaries.

- sendKeysChunkedToTarget now skips the 50ms inter-chunk delay for chunks
  that already end in a newline (the common case for large multi-line
  payloads, since splitIntoChunks prefers newline boundaries) — that line
  was already flushed through the line discipline, so pacing it further
  only adds latency. Only the no-newline hard-split fallback, which the
  delay was meant to pace, still sleeps. This avoids a ~4x latency and
  subprocess-count regression on large newline-containing transfers (e.g.
  the 500KB TUI-paste path in internal/ui/home.go) versus the old
  4096-byte chunkSize.

- Renamed issue1793_canonical_chunk_test.go to tmux_chunk_size_test.go and
  rewrote its header/test-name comments: it only proves an argv-shape
  invariant (chunk sizes, byte-for-byte reassembly, separate trailing
  Enter) via a stubbed tmux exec seam, not real pane/tty delivery, and
  cannot detect whether the underlying phantom-send mechanism is actually
  fixed. Real pane-level delivery is already covered by
  internal/integration/conductor_test.go's TestConductor_ChunkedSendDelivery.

- Corrected stale "4096-byte chunk threshold" comments in conductor_test.go
  to reference the current 1023-byte chunkSize.

This remains a correctness stopgap, not a proven fix for the reported
mechanism; the strategic fix (mailbox/receipt delivery model that proves
turn-submission instead of inferring it from transport success) is tracked
separately per #1793.

Committed by Ashesh Goplani
